MiCodigo/HF_PC3/requirements.txt
scikit-learn
fastapi
numpy
opencv-python
setuptools
minisom
requests
pydantic
opencv-python-headless
uvicorn[standard]
python-multipart
pillow
matplotlib